- Changed cell width to 64-bits. - Fixed a few bugs associated with 32-bit to 64-bit cell transition. (Interpreter) - Added basic double implementation (MUST WORK MORE ON IT!) (TODO: JIT IMPLEMENTATION)

+++ compiler/compile-options.h
@@ -36,7 +36,7 @@ struct CompileOptions {
     bool require_newdecls = false;
     bool warnings_are_errors = false;
     bool use_stderr = false;
-    int pragma_dynamic = 0;
+    int64_t pragma_dynamic = 0;
     int ctrlchar_org = CTRL_CHAR;
     int compression = 9;
     bool show_includes = false;
